Rejig matrix vector multiplication for general types
Should now work on more than just Vec and Mat, for example you could
change between RGB colour spaces using Mat*RGB multiplication, and get
back an RGB as the output.

The major implementation change here is to use construct_similar rather
than assuming the output vector type, and to just unroll the product
entirely rather than emitting a bunch of calls to bilindot().  As a
result, a bunch of bilindot stuff can be removed since we don't need to
extract rows of the matrix as tuples.  (I guess this is good, though it
may result in rather some code bloat for larger FSAs.  Can't really be
solved properly until we get setindex() on a Ref or some such thing.)