Reduced code generation in mat*vec and mat*mat multiply
Heuristic thresholds to limit the size of generated expressions in
mat*mat and mat*vec loop unrolling.  The generated machine code still
isn't necessarily good here, but this improves compile times to
something acceptable.

An alternative approach could be to explicitly unpack the matrix rows
and columns onto the stack into a set of local variables.